refactor: add tests and removed unnecessary annotations to DisplayPermissionResponse (SDKCF-6394)
