+++ inappmessaging/src/main/java/com/rakuten/tech/mobile/inappmessaging/runtime/api/MessageMixerRetrofitService.kt
@@ -3,7 +3,7 @@ package com.rakuten.tech.mobile.inappmessaging.runtime.api
 import com.rakuten.tech.mobile.inappmessaging.runtime.data.requests.DisplayPermissionRequest
 import com.rakuten.tech.mobile.inappmessaging.runtime.data.requests.ImpressionRequest
 import com.rakuten.tech.mobile.inappmessaging.runtime.data.requests.PingRequest
-import com.rakuten.tech.mobile.inappmessaging.runtime.data.responses.displaypermission.DisplayPermissionResponse
+import com.rakuten.tech.mobile.inappmessaging.runtime.data.responses.DisplayPermissionResponse
 import com.rakuten.tech.mobile.inappmessaging.runtime.data.responses.ping.MessageMixerResponse
 import okhttp3.ResponseBody
 import retrofit2.Call
