+++ inappmessaging/src/main/java/com/rakuten/tech/mobile/inappmessaging/runtime/data/responses/DisplayPermissionResponse.kt
@@ -1,0 +1,10 @@
+package com.rakuten.tech.mobile.inappmessaging.runtime.data.responses
+
+import androidx.annotation.RestrictTo
+import com.google.gson.annotations.SerializedName
+
+@RestrictTo(RestrictTo.Scope.LIBRARY)
+internal data class DisplayPermissionResponse(
+    val display: Boolean,
+    @SerializedName("performPing") val shouldPing: Boolean,
+)
